requirements.txt
requests==2.18.1
PyJWT==1.5.2
pytest==3.1.3
docopt==0.6.2
colorlog==2.10.0
